Fixes issue 857.
Big thanks goes to taoshaof.. for pointing this out. 

The fix is to re raise the call screen for all states.
A pass has to be done in call screens to ensure this problem can not also appear (possibly so). But as the entire call screen will be refactored tomorrow I'll do that tomorrow :).



git-svn-id: http://csipsimple.googlecode.com/svn/trunk@775 9f815046-5998-e9c0-b7e2-ac03a23edfa4